Isolate solr updater date-related functions
